ClockP: fix us/ms <-> ticks conversion errors
Calculations based on the clock tick period produce series loss in
precision (~17250 ppm) when the system tick is based on the LF crystal.
This is inacceptable for most applications especially in the RF area and
leads to fatal timing errors there (e.g. the radio timer oversleeping
systematically during power down).

This change improves precision by rounding as late as possible. To this
end calculations are based on the precise tick frequency rather than
rounded tick periods.

While rounding later improves precision it will not completely remove
rounding error in most cases. Therefore an appropriate rounding approach
must be chosen for each individual use case.

Several helper functions with well-defined rounding semantics are
introduced and applied throughout the HAL to improve precision as much
as possible in the integers domain.

Signed-off-by: Florian Grandel <fgrandel@code-for-humans.de>