hal: drivers: radio: Port radio driver for F3 devices to Zephyr
Port RCL (radio control layer), the radio driver for TI F3 devices to
Zephyr. This commit includes the source files of RCL, binary images of
modem firmwares, and pre-generated radio settings for adc noise and BLE5.
NOTE: currently, the radio settings for BLE5 are compiled as part of the
driver library by default, which is a workaround for now.

TI's BLE Stack controller requires the radio configuration C structs
without the PHY abbreviation appended as a suffix. This commit removes
the PHY abbreviation by disabling the "use_phy_abbreviation" when
generating the radio settings in SmartRF Studio 8.

Signed-off-by: Min Xu <m-xu@ti.com>