hal: add startup code for TMS570LC43
Adds soc init code including MPU configuration and PLL errata
handling for the TI Hercules family SoC TMS570LC43xx. Portions
of this code were taken or adapter from Halcogen code generator.

This HAL is not really complete, as the currently implemented
drivers do not depend on Halcogen and so were not needed.

Signed-off-by: Tavish Naruka <t-naruka@ispace-inc.com>